refactored drivers module into a sub-package
